Move DiscordClient to the presentation layer
